Add test 에서 FromContainer test 분리 후 컴포넌트 구현
