+++ src/App.jsx
@@ -1,31 +1,11 @@
-import { useSelector, useDispatch } from 'react-redux';
+import { useSelector } from 'react-redux';
 
-import {
-  updateInformation,
-  addRestaurant,
-} from './actions';
+import FormContainer from './FormContainer';
 
 export default function App() {
-  const { information, restaurants } = useSelector((state) => ({
-    information: state.information,
+  const { restaurants } = useSelector((state) => ({
     restaurants: state.restaurants,
   }));
-
-  const { name, classification, address } = information;
-
-  const dispatch = useDispatch();
-
-  function handleChange(event) {
-    const { id, value } = event.target;
-
-    dispatch(updateInformation(id, value));
-  }
-
-  function handleSubmit(event) {
-    event.preventDefault();
-
-    dispatch(addRestaurant());
-  }
 
   return (
     <>
@@ -37,12 +17,7 @@ export default function App() {
           </li>
         ))}
       </ul>
-      <form onSubmit={handleSubmit}>
-        <input id="name" type="text" placeholder="이름" value={name} onChange={handleChange} />
-        <input id="classification" type="text" placeholder="분류" value={classification} onChange={handleChange} />
-        <input id="address" type="text" placeholder="주소" value={address} onChange={handleChange} />
-        <button type="submit">등록</button>
-      </form>
+      <FormContainer />
     </>
   );
 }
